Updated encoder data types to reflect the amount of data stored in them (x64 processors default to 32 bits/integer).

+++ src/wr_control_drive_arm/src/ArmMotor.cpp
@@ -2,8 +2,8 @@
 
 #define Std_UInt32 std_msgs::UInt32::ConstPtr&
 
-unsigned long int const ArmMotor::COUNTS_PER_ROTATION = UINT32_MAX;
-unsigned long int const ArmMotor::ENCODER_BOUNDS[2] = {0, ArmMotor::COUNTS_PER_ROTATION};
+unsigned int const ArmMotor::COUNTS_PER_ROTATION = UINT32_MAX;
+unsigned int const ArmMotor::ENCODER_BOUNDS[2] = {0, ArmMotor::COUNTS_PER_ROTATION};
 
 template<class T> T ArmMotor::corrMod(T i, T j){
     int dir = -((i-j>0)-(i-j<0));
@@ -11,7 +11,7 @@ template<class T> T ArmMotor::corrMod(T i, T j){
     return i;
 }
 
-unsigned long int ArmMotor::radToEnc(double rads){
+unsigned int ArmMotor::radToEnc(double rads){
     return ArmMotor::COUNTS_PER_ROTATION*ArmMotor::corrMod(rads,2 * M_PI)/(2 * M_PI);
 }
 
@@ -29,7 +29,7 @@ void ArmMotor::storeEncoderVals(const Std_UInt32 msg){
 
 ArmMotor::ArmMotor(){}
 
-ArmMotor::ArmMotor(std::string motorName, unsigned long int controllerID, unsigned long int motorID, ros::NodeHandle* n){
+ArmMotor::ArmMotor(std::string motorName, unsigned int controllerID, unsigned int motorID, ros::NodeHandle* n){
     if(controllerID > 3) throw ((std::string)"Controller ID ") + std::to_string(controllerID) + "is only valid on [0,3]";
     if(motorID > 1) throw ((std::string)"Motor ID ") + std::to_string(motorID) + "is only valid on [0,1]";
     
@@ -46,26 +46,26 @@ ArmMotor::ArmMotor(std::string motorName, unsigned long int controllerID, unsign
 
 // ~ArmMotor();
 
-unsigned long int ArmMotor::getEncoderCounts(){
+unsigned int ArmMotor::getEncoderCounts(){
     return this->encoderVal;
 }
 
 // void resetEncoder();
 
-void ArmMotor::runToTarget(unsigned long int targetCounts, float power){
+void ArmMotor::runToTarget(unsigned int targetCounts, float power){
     this->runToTarget(targetCounts, power, false);
 }
 
-bool ArmMotor::hasReachedTarget(unsigned long int targetCounts, unsigned long int tolerance){
-    unsigned long int lBound = ((targetCounts - tolerance) % ArmMotor::ENCODER_BOUNDS[1] < 0 ? 2 * targetCounts - tolerance : targetCounts - tolerance) % ArmMotor::ENCODER_BOUNDS[1];
-    unsigned long int uBound = (targetCounts + tolerance) % ArmMotor::ENCODER_BOUNDS[1];
+bool ArmMotor::hasReachedTarget(unsigned int targetCounts, unsigned int tolerance){
+    unsigned int lBound = ((targetCounts - tolerance) % ArmMotor::ENCODER_BOUNDS[1] < 0 ? 2 * targetCounts - tolerance : targetCounts - tolerance) % ArmMotor::ENCODER_BOUNDS[1];
+    unsigned int uBound = (targetCounts + tolerance) % ArmMotor::ENCODER_BOUNDS[1];
     if(lBound < uBound)
         return this->getEncoderCounts() <= uBound && this->getEncoderCounts() >=lBound;
     else
         return this->getEncoderCounts() <= uBound || this->getEncoderCounts() >=lBound;
 }
 
-bool ArmMotor::hasReachedTarget(unsigned long int targetCounts){
+bool ArmMotor::hasReachedTarget(unsigned int targetCounts){
     return ArmMotor::hasReachedTarget(targetCounts, 1200000);
 }
 
@@ -82,10 +82,10 @@ void ArmMotor::setPower(float power){
     this->currState = power == 0.f ? MotorState::STOP : MotorState::MOVING;
 }
 
-void ArmMotor::runToTarget(unsigned long int targetCounts, float power, bool block){
+void ArmMotor::runToTarget(unsigned int targetCounts, float power, bool block){
     if(this->getMotorState() != MotorState::RUN_TO_TARGET && !this->hasReachedTarget(targetCounts)){
-        long long int direction = targetCounts - this->getEncoderCounts();
-        power = abs(power) * (corrMod(direction, ((long long int)ArmMotor::COUNTS_PER_ROTATION)) < corrMod(-direction, ((long long int)ArmMotor::COUNTS_PER_ROTATION)) ? 1 : -1);
+        long int direction = targetCounts - this->getEncoderCounts();
+        power = abs(power) * (corrMod(direction, ((long int)ArmMotor::COUNTS_PER_ROTATION)) < corrMod(-direction, ((long int)ArmMotor::COUNTS_PER_ROTATION)) ? 1 : -1);
         this->setPower(power);
         this->currState = MotorState::RUN_TO_TARGET;
     }
